refactor: refactor HeaderDashboard component to be shared
