+++ app/[locale]/components/templates/Variant.tsx
@@ -1,10 +1,10 @@
-import HeaderVariant from '@organisms/variant/HeaderVariant';
 import MainVariant from '@organisms/variant/MainVariant';
+import HeaderDashboard from '@organisms/shared/HeaderDashboard';
 
 export default function VariantTemplate() {
   return (
     <div className="bg-background flex min-h-screen flex-col">
-      <HeaderVariant />
+      <HeaderDashboard />
       <MainVariant />
     </div>
   );
